fix for #365
git-svn-id: https://svn.kvirc.de/svn/trunk/kvirc@3173 17fca916-40b9-46aa-a4ea-0a15b648b75c

+++ src/modules/dcc/send.h
@@ -199,7 +199,7 @@ private:
 	QTimer                 * m_pResumeTimer; // used to signal resume timeout
 public:
 	bool resumeAccepted(const char * filename,const char * port,const char *szZeroPortTag);
-	bool doResume(const char * filename,const char * port,unsigned int filePos);
+	bool doResume(const char * filename,const char * port,unsigned long filePos);
 
 	static void init();
 	static void done();
@@ -207,7 +207,7 @@ public:
 	static KviDccFileTransfer * nonFailedTransferWithLocalFileName(const QString &szLocalFileName);
 	static unsigned int transferCount();
 	static bool handleResumeAccepted(const char * filename,const char * port,const char * szZeroPortTag);
-	static bool handleResumeRequest(const char * filename,const char * port,unsigned int filePos);
+	static bool handleResumeRequest(const char * filename,const char * port,unsigned long filePos);
 
 	virtual bool event(QEvent *e);
